Move handler and data types to config-site crate
