+++ src/board/wifi/mod.rs
@@ -1,2 +1,1 @@
 pub mod driver;
-pub mod network;
